+++ app/src/main/java/io/vn/nguyenduck/blocktopograph/activity/navigation/SettingFragment.java
@@ -13,8 +13,6 @@ import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
 import java.util.ArrayList;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
 
 import io.vn.nguyenduck.blocktopograph.R;
 import io.vn.nguyenduck.blocktopograph.activity.subview.WorldScannerFolderView;
@@ -23,7 +21,6 @@ import io.vn.nguyenduck.blocktopograph.setting.SettingManager;
 public class SettingFragment extends Fragment {
 
     private SettingAdapter adapter;
-    private final ExecutorService EXECUTOR_SERVICE = Executors.newWorkStealingPool(1);
     private final SettingManager MANAGER = SettingManager.getInstance();
 
     @Override
@@ -40,39 +37,34 @@ public class SettingFragment extends Fragment {
         return v;
     }
 
-    private void addIntoExecutor(Runnable task) {
-        if (EXECUTOR_SERVICE.isShutdown()) return;
-        EXECUTOR_SERVICE.execute(task);
-    }
-
     @Override
     public void onAttach(@NonNull Context context) {
         super.onAttach(context);
-        addIntoExecutor(MANAGER::load);
+        SettingManager.forceLoad();
     }
 
     @Override
     public void onDetach() {
         super.onDetach();
-        addIntoExecutor(MANAGER::save);
+        SettingManager.forceSave();
     }
 
     @Override
     public void onStart() {
         super.onStart();
-        addIntoExecutor(MANAGER::load);
+        SettingManager.forceLoad();
     }
 
     @Override
     public void onPause() {
         super.onPause();
-        addIntoExecutor(MANAGER::save);
+        SettingManager.forceSave();
     }
 
     @Override
     public void onDestroy() {
         super.onDestroy();
-        EXECUTOR_SERVICE.shutdown();
+        SettingManager.shutdown();
     }
 
     private static class SettingAdapter extends BaseAdapter {