+++ release_notes.md
@@ -34,5 +34,7 @@
 
 ### 01Nov24
 
-- Completed world scan folders system can add, edit, remove, save and load settings
-- Fix some error make app crash
+- Completed world scan folders system:
+    - Added ability to add, edit, and remove scan folders
+    - Implemented settings persistence (save/load)
+- Fixed world loader update item make app crash